gen_parser builds now and the pegged update seems to work and even solve a previous blocker!  Build is still broken because the WIP code from before is probably broken.

+++ gen_parser.d
@@ -2,8 +2,7 @@ import std.file;
 import std.stdio;
 import std.conv;
 
-//import pegged.examples.dparser;
-//import pegged.examples.c;
+import pegged.examples.c;
 import pegged.grammar;
 
 import grammars.pmlgrammar;
@@ -12,7 +11,7 @@ void main()
 {
 	if ( !std.file.exists("generated") )
 		std.file.mkdir("generated");
-	/+if ( std.file.exists("generated/dparser.d") )
+	if ( std.file.exists("generated/dparser.d") )
 		std.file.remove("generated/dparser.d");
 	
 	std.file.write("generated/dparser.d",
@@ -20,7 +19,7 @@ void main()
 			"module generated.dparser;\n"~
 			"import pegged.grammar;\n"~
 			"\n"~grammar(Cgrammar)));
-	+/
+	
 	std.file.write("generated/pml.d",
         to!string(
             "module generated.pml;\n"~
